Separate formatter from TextUi
